# docs/swagger.yaml
basePath: /api
definitions:
  customerror.Error:
    properties:
      errors:
        additionalProperties: true
        type: object
    type: object
  user.User:
    properties:
      bio:
        description: '`bson:"bio"`'
        type: string
      email:
        type: string
      id:
        type: string
      password:
        type: string
      roles:
        items:
          type: integer
        type: array
      username:
        type: string
    type: object
  user.userLoginRequest:
    properties:
      user:
        properties:
          email:
            type: string
          password:
            type: string
        required:
        - email
        - password
        type: object
    type: object
  user.userRegisterRequest:
    properties:
      user:
        properties:
          email:
            type: string
          password:
            type: string
          username:
            type: string
        required:
        - email
        - password
        - username
        type: object
    type: object
  user.userResponse:
    properties:
      user:
        properties:
          bio:
            type: string
          email:
            type: string
          id:
            type: string
          image:
            type: string
          roles:
            items:
              type: integer
            type: array
          token:
            type: string
          username:
            type: string
        type: object
    type: object
  user.userUpdateRequest:
    properties:
      user:
        properties:
          bio:
            type: string
          email:
            type: string
          password:
            type: string
          username:
            type: string
        type: object
    type: object
host: 127.0.0.1:8585
info:
  contact: {}
  description: Conduit API
  title: Conduit API
  version: "1.3"
paths:
  /user:
    get:
      consumes:
      - application/json
      description: Users list for Admin
      operationId: list-user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/user.User'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/customerror.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/customerror.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/customerror.Error'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/customerror.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/customerror.Error'
      security:
      - ApiKeyAuth: []
      summary: Users list
      tags:
      - user
    post:
      consumes:
      - application/json
      description: Add new user
      operationId: add-user
      parameters:
      - description: User info for registration
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user.userRegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/user.userResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/customerror.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/customerror.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/customerror.Error'
      security:
      - ApiKeyAuth: []
      summary: Add new user
      tags:
      - user
    put:
      consumes:
      - application/json
      description: Update user information for current user
      operationId: update-profile
      parameters:
      - description: User details to update. At least **one** field is required.
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user.userUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.userResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/customerror.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/customerror.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/customerror.Error'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/customerror.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/customerror.Error'
      security:
      - ApiKeyAuth: []
      summary: Update profile
      tags:
      - user
  /user/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a user
      operationId: delete-user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/user.userResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/customerror.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/customerror.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/customerror.Error'
      security:
      - ApiKeyAuth: []
      summary: Delete a user
      tags:
      - user
    put:
      consumes:
      - application/json
      description: Update user information for current user
      operationId: update-user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: User details to update. At least **one** field is required.
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user.userUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.userResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/customerror.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/customerror.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/customerror.Error'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/customerror.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/customerror.Error'
      security:
      - ApiKeyAuth: []
      summary: Update current user
      tags:
      - user
  /user/info:
    get:
      consumes:
      - application/json
      description: Gets the currently logged-in user
      operationId: current-user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.userResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/customerror.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/customerror.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/customerror.Error'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/customerror.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/customerror.Error'
      security:
      - ApiKeyAuth: []
      summary: Get the current user
      tags:
      - user
  /users:
    post:
      consumes:
      - application/json
      description: Register a new user
      operationId: sign-up
      parameters:
      - description: User info for registration
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user.userRegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/user.userResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/customerror.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/customerror.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/customerror.Error'
      summary: Register a new user
      tags:
      - user
  /users/login:
    post:
      consumes:
      - application/json
      description: Login for existing user
      operationId: login
      parameters:
      - description: Credentials to use
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/user.userLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user.userResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/customerror.Error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/customerror.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/customerror.Error'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/customerror.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/customerror.Error'
      summary: Login for existing user
      tags:
      - user
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"